remove stage sorting
