+++ packages/frontend/src/pages/scaling/activity/components/table/ScalingActivityTable.tsx
@@ -1,7 +1,6 @@
 import { getCoreRowModel, getSortedRowModel } from '@tanstack/react-table'
 import { useMemo } from 'react'
 import { BasicTable } from '~/components/table/BasicTable'
-import { RollupsTable } from '~/components/table/RollupsTable'
 import { useTableSorting } from '~/components/table/sorting/TableSortingContext'
 import { useTable } from '~/hooks/useTable'
 import type { ScalingActivityEntry } from '~/server/features/scaling/activity/getScalingActivityEntries'
@@ -12,11 +11,10 @@ import { getScalingActivityColumns } from './columns'
 
 interface Props {
   entries: ScalingActivityEntry[]
-  rollups?: boolean
   notReviewed?: boolean
 }
 
-export function ScalingActivityTable({ entries, rollups, notReviewed }: Props) {
+export function ScalingActivityTable({ entries, notReviewed }: Props) {
   const { metric } = useActivityMetricContext()
   const { sorting, setSorting } = useTableSorting()
 
@@ -45,9 +43,7 @@ export function ScalingActivityTable({ entries, rollups, notReviewed }: Props) {
       },
     },
   })
-  return rollups ? (
-    <RollupsTable table={table} />
-  ) : (
+  return (
     <BasicTable
       table={table}
       rowColoringMode={notReviewed ? 'ignore-colors' : undefined}
