+++ packages/frontend/src/pages/scaling/costs/components/table/ScalingCostsTable.tsx
@@ -1,7 +1,6 @@
 import { getCoreRowModel, getSortedRowModel } from '@tanstack/react-table'
 import { useEffect, useMemo } from 'react'
 import { BasicTable } from '~/components/table/BasicTable'
-import { RollupsTable } from '~/components/table/RollupsTable'
 import { useTableSorting } from '~/components/table/sorting/TableSortingContext'
 import { useTable } from '~/hooks/useTable'
 import type { CostsTableData } from '~/server/features/scaling/costs/getCostsTableData'
@@ -17,10 +16,9 @@ import { getScalingCostsColumns } from './columns'
 
 interface Props {
   entries: ScalingCostsEntry[]
-  rollups?: boolean
 }
 
-export function ScalingCostsTable({ entries, rollups }: Props) {
+export function ScalingCostsTable({ entries }: Props) {
   const { range } = useCostsTimeRangeContext()
   const { unit } = useCostsUnitContext()
   const { metric } = useCostsMetricContext()
@@ -61,7 +59,7 @@ export function ScalingCostsTable({ entries, rollups }: Props) {
     }
   }, [metric, setSorting, table])
 
-  return rollups ? <RollupsTable table={table} /> : <BasicTable table={table} />
+  return <BasicTable table={table} />
 }
 
 function mapToTableEntry(
