+++ packages/frontend/src/pages/scaling/data-availability/components/table/ScalingDaTable.tsx
@@ -1,6 +1,5 @@
 import { getCoreRowModel, getSortedRowModel } from '@tanstack/react-table'
 import { BasicTable } from '~/components/table/BasicTable'
-import { RollupsTable } from '~/components/table/RollupsTable'
 import { useTableSorting } from '~/components/table/sorting/TableSortingContext'
 import { useTable } from '~/hooks/useTable'
 import type { ScalingDaEntry } from '~/server/features/scaling/data-availability/getScalingDaEntries'
@@ -8,11 +7,10 @@ import { getScalingDataAvailabilityColumns } from './columns'
 
 interface Props {
   entries: ScalingDaEntry[]
-  rollups?: boolean
   hideType?: boolean
 }
 
-export function ScalingDaTable({ entries, rollups, hideType }: Props) {
+export function ScalingDaTable({ entries, hideType }: Props) {
   const { sorting, setSorting } = useTableSorting()
   const table = useTable({
     data: entries,
@@ -31,5 +29,5 @@ export function ScalingDaTable({ entries, rollups, hideType }: Props) {
     },
   })
 
-  return rollups ? <RollupsTable table={table} /> : <BasicTable table={table} />
+  return <BasicTable table={table} />
 }
