+++ packages/frontend/src/pages/scaling/risk/components/ScalingRiskTables.tsx
@@ -68,7 +68,7 @@ export function ScalingRiskTables(props: Props) {
         <TableSortingProvider initialSort={initialSort}>
           <DirectoryTabsContent value="rollups">
             <RollupsInfo />
-            <ScalingRiskTable entries={entries.rollups} rollups />
+            <ScalingRiskTable entries={entries.rollups} />
           </DirectoryTabsContent>
         </TableSortingProvider>
         <TableSortingProvider initialSort={initialSort}>
