+++ packages/frontend/src/pages/scaling/tvs/components/table/ScalingTvsTable.tsx
@@ -1,7 +1,6 @@
 import { getCoreRowModel, getSortedRowModel } from '@tanstack/react-table'
 import { useMemo } from 'react'
 import { BasicTable } from '~/components/table/BasicTable'
-import { RollupsTable } from '~/components/table/RollupsTable'
 import { useTableSorting } from '~/components/table/sorting/TableSortingContext'
 import { useTable } from '~/hooks/useTable'
 import type { ScalingTvsEntry } from '~/server/features/scaling/tvs/getScalingTvsEntries'
@@ -11,11 +10,10 @@ import { getScalingTvsColumns } from './columns'
 
 interface Props {
   entries: ScalingTvsEntry[]
-  rollups?: boolean
   notReviewed?: boolean
 }
 
-export function ScalingTvsTable({ entries, rollups, notReviewed }: Props) {
+export function ScalingTvsTable({ entries, notReviewed }: Props) {
   const { excludeAssociatedTokens } = useScalingAssociatedTokensContext()
   const { sorting, setSorting } = useTableSorting()
 
@@ -45,9 +43,7 @@ export function ScalingTvsTable({ entries, rollups, notReviewed }: Props) {
     },
   })
 
-  return rollups ? (
-    <RollupsTable table={table} />
-  ) : (
+  return (
     <BasicTable
       table={table}
       insideMainPageCard
