+++ packages/frontend/src/server/features/scaling/summary/getScalingSummaryEntries.ts
@@ -24,8 +24,8 @@ import { getCommonScalingEntry } from '../getCommonScalingEntry'
 import { getApprovedOngoingAnomalies } from '../liveness/getApprovedOngoingAnomalies'
 import type { ProjectSevenDayTvsBreakdown } from '../tvs/get7dTvsBreakdown'
 import { get7dTvsBreakdown } from '../tvs/get7dTvsBreakdown'
+import { compareTvs } from '../tvs/utils/compareTvs'
 import { getAssociatedTokenWarning } from '../tvs/utils/getAssociatedTokenWarning'
-import { compareStageAndTvs } from '../utils/compareStageAndTvs'
 
 export async function getScalingSummaryEntries() {
   const projects = await ps.getProjects({
@@ -57,7 +57,7 @@ export async function getScalingSummaryEntries() {
         !!projectsOngoingAnomalies[project.id.toString()],
       ),
     )
-    .sort(compareStageAndTvs)
+    .sort(compareTvs)
 
   return groupByScalingTabs(entries)
 }
